Fixing post request bug for lists
